requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.3.0
python-dotenv
pydantic[email]
pydantic-settings
python-multipart
emails
